requirements.txt
opencv-python
opencv-contrib-python
numpy
matplotlib
mediapipe
uuid
moviepy
python-dotenv